Changing resume to use modencv
Changing resume style to make it far easier to modify the resume, the
frigerri-cv didn't allow page breaks in sections making formatting
incredibly difficult. Moderncv gets around these limitations.

There is also the addition of the cover letter which is in the same
style as the resume. The preamble, the cover letter and the resume have
been separated into separate files, with the cover letters found in the
letters directory, one for each job application. This allows me to keep
track of all the job applications and have the cover letters in one
place. I am adding the job requirements as comments to the cover
letter for later reference and to ensure all the criteria are met.