Remove bibliography components
The tectonic update now automatically runs biber, which in this case is
not added to the environment so the build fails. By removing all the
references to the biliography elements, the .bcf file which tectonic
uses to check whether to run biber is not generated and so we no longer
get the build failure.